# api/openapi.yaml
components:
    schemas:
        APIListResponse:
            properties:
                items:
                    items:
                        $ref: '#/components/schemas/APIResponse'
                    type: array
                kind:
                    type: string
                total:
                    type: integer
            required:
                - kind
                - items
                - total
            type: object
...
        APIResponse:
            properties:
                kind:
                    $ref: '#/components/schemas/ResourceKind'
                metadata:
                    $ref: '#/components/schemas/ResourceMeta'
                spec:
                    $ref: '#/components/schemas/APISpec'
                status:
                    $ref: '#/components/schemas/APIStatus'
            required:
                - kind
                - metadata
                - spec
            type: object
...
        ApplyRequest:
            properties:
                resources:
                    items:
                        additionalProperties: true
                        type: object
                    type: array
            required:
                - resources
            type: object
...
        DeleteResponse:
            properties:
                message:
                    type: string
            required:
                - message
            type: object
...
        ErrorResponse:
            properties:
                code:
                    format: int64
                    type: integer
                details:
                    additionalProperties:
                        type: string
                    type: object
                error:
                    type: string
            required:
                - error
                - code
            type: object
...
info:
    description: Declarative Envoy xDS control plane with reconciliation-based architecture. Resources (Gateways, Listeners, Environments, APIs, Deployments) are declared via this REST API and stored in a desired-state store.
    license:
        name: Apache 2.0
    title: FlowC Control Plane API
    version: 3.0.0
openapi: 3.0.3
paths:
    /api/v1/apis:
        get:
            operationId: listAPIs
            parameters:
                - description: Label filter (comma-separated key=value pairs)
                  in: query
                  name: labels
                  schema:
                    type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/APIListResponse'
                    description: List of apis
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Error response
            summary: List apis
            tags:
                - APIs
    /api/v1/apis/{name}:
        delete:
            operationId: deleteAPI
            parameters:
                - description: Resource revision for optimistic concurrency control
                  in: header
                  name: If-Match
                  schema:
                    format: int64
                    type: integer
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/DeleteResponse'
                    description: Resource deleted
                "404":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Error response
                "409":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Error response
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Error response
            summary: Delete a API
            tags:
                - APIs
        get:
            operationId: getAPI
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/APIResponse'
                    description: The API resource
                "404":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Error response
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Error response
            summary: Get a API
            tags:
                - APIs
        parameters:
            - description: Resource name
              in: path
              name: name
              required: true
              schema:
                type: string
        put:
            operationId: putAPI
            parameters:
                - description: Resource revision for optimistic concurrency control
                  in: header
                  name: If-Match
                  schema:
                    format: int64
                    type: integer
                - description: Identifier of the managing entity for ownership tracking
                  in: header
                  name: X-Managed-By
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/APIPutRequest'
                required: true
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/APIResponse'
                    description: Resource updated
                "201":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/APIResponse'
                    description: Resource created
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Error response
                "409":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Error response
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Error response
            summary: Create or update a API
            tags:
                - APIs
    /api/v1/apply:
        post:
            description: Creates or updates multiple resources in a single request.
            operationId: applyResources
            parameters:
                - description: Identifier of the managing entity for ownership tracking
                  in: header
                  name: X-Managed-By
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/ApplyRequest'
                required: true
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ApplyResult'
                    description: Apply results
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Error response
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Error response
            summary: Bulk apply resources
            tags:
                - Apply
    /api/v1/deployments:
        get:
            operationId: listDeployments
            parameters:
                - description: Label filter (comma-separated key=value pairs)
                  in: query
                  name: labels
                  schema:
                    type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/DeploymentListResponse'
                    description: List of deployments
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Error response
            summary: List deployments
            tags:
                - Deployments
    /api/v1/deployments/{name}:
        delete:
            operationId: deleteDeployment
            parameters:
                - description: Resource revision for optimistic concurrency control
                  in: header
                  name: If-Match
                  schema:
                    format: int64
                    type: integer
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/DeleteResponse'
                    description: Resource deleted
                "404":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Error response
                "409":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Error response
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Error response
            summary: Delete a Deployment
            tags:
                - Deployments
        get:
            operationId: getDeployment
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/DeploymentResponse'
                    description: The Deployment resource
                "404":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Error response
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Error response
            summary: Get a Deployment
            tags:
                - Deployments
        parameters:
            - description: Resource name
              in: path
              name: name
              required: true
              schema:
                type: string
        put:
            operationId: putDeployment
            parameters:
                - description: Resource revision for optimistic concurrency control
                  in: header
                  name: If-Match
                  schema:
                    format: int64
                    type: integer
                - description: Identifier of the managing entity for ownership tracking
                  in: header
                  name: X-Managed-By
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/DeploymentPutRequest'
                required: true
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/DeploymentResponse'
                    description: Resource updated
                "201":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/DeploymentResponse'
                    description: Resource created
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Error response
                "409":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Error response
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Error response
            summary: Create or update a Deployment
            tags:
                - Deployments
    /api/v1/environments:
        get:
            operationId: listEnvironments
            parameters:
                - description: Label filter (comma-separated key=value pairs)
                  in: query
                  name: labels
                  schema:
                    type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/EnvironmentListResponse'
                    description: List of environments
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Error response
            summary: List environments
            tags:
                - Environments
    /api/v1/environments/{name}:
        delete:
            operationId: deleteEnvironment
            parameters:
                - description: Resource revision for optimistic concurrency control
                  in: header
                  name: If-Match
                  schema:
                    format: int64
                    type: integer
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/DeleteResponse'
                    description: Resource deleted
                "404":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Error response
                "409":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Error response
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Error response
            summary: Delete a Environment
            tags:
                - Environments
        get:
            operationId: getEnvironment
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/EnvironmentResponse'
                    description: The Environment resource
                "404":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Error response
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Error response
            summary: Get a Environment
            tags:
                - Environments
        parameters:
            - description: Resource name
              in: path
              name: name
              required: true
              schema:
                type: string
        put:
            operationId: putEnvironment
            parameters:
                - description: Resource revision for optimistic concurrency control
                  in: header
                  name: If-Match
                  schema:
                    format: int64
                    type: integer
                - description: Identifier of the managing entity for ownership tracking
                  in: header
                  name: X-Managed-By
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/EnvironmentPutRequest'
                required: true
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/EnvironmentResponse'
                    description: Resource updated
                "201":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/EnvironmentResponse'
                    description: Resource created
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Error response
                "409":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Error response
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Error response
            summary: Create or update a Environment
            tags:
                - Environments
    /api/v1/gatewayprofiles:
        get:
            operationId: listGatewayProfiles
            parameters:
                - description: Label filter (comma-separated key=value pairs)
                  in: query
                  name: labels
                  schema:
                    type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/GatewayProfileListResponse'
                    description: List of gatewayprofiles
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Error response
            summary: List gatewayprofiles
            tags:
                - GatewayProfiles
    /api/v1/gatewayprofiles/{name}:
        delete:
            operationId: deleteGatewayProfile
            parameters:
                - description: Resource revision for optimistic concurrency control
                  in: header
                  name: If-Match
                  schema:
                    format: int64
                    type: integer
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/DeleteResponse'
                    description: Resource deleted
                "404":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Error response
                "409":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Error response
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Error response
            summary: Delete a GatewayProfile
            tags:
                - GatewayProfiles
        get:
            operationId: getGatewayProfile
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/GatewayProfileResponse'
                    description: The GatewayProfile resource
                "404":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Error response
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Error response
            summary: Get a GatewayProfile
            tags:
                - GatewayProfiles
        parameters:
            - description: Resource name
              in: path
              name: name
              required: true
              schema:
                type: string
        put:
            operationId: putGatewayProfile
            parameters:
                - description: Resource revision for optimistic concurrency control
                  in: header
                  name: If-Match
                  schema:
                    format: int64
                    type: integer
                - description: Identifier of the managing entity for ownership tracking
                  in: header
                  name: X-Managed-By
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/GatewayProfilePutRequest'
                required: true
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/GatewayProfileResponse'
                    description: Resource updated
                "201":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/GatewayProfileResponse'
                    description: Resource created
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Error response
                "409":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Error response
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Error response
            summary: Create or update a GatewayProfile
            tags:
                - GatewayProfiles
    /api/v1/gateways:
        get:
            operationId: listGateways
            parameters:
                - description: Label filter (comma-separated key=value pairs)
                  in: query
                  name: labels
                  schema:
                    type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/GatewayListResponse'
                    description: List of gateways
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Error response
            summary: List gateways
            tags:
                - Gateways
...
    /api/v1/gateways/{name}/bootstrap:
        get:
            description: Generates an Envoy bootstrap YAML for the gateway, using the referenced profile's settings (admin port, xDS cluster, etc.).
            operationId: getGatewayBootstrap
            responses:
                "200":
                    content:
                        application/x-yaml:
                            schema:
                                type: string
                    description: Envoy bootstrap YAML
                "404":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Error response
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Error response
            summary: Generate Envoy bootstrap configuration
            tags:
                - Bootstrap
        parameters:
            - description: Gateway name
              in: path
              name: name
              required: true
              schema:
                type: string
...
servers:
    - description: Local development server
      url: http://localhost:8080
...